Simplify concatenation of formatted strings
Instead of keeping track of offset into and length of buffer,
keep track of pointers to next input and to end of buffer.
This means one less subtraction in every call to g_snprintf()
and occationally statements that don't have to span several lines.

The --tlsverify flag was also moved from tls.c to the main file.